Updated Keypair & keypair_test based on comments
