+++ lib/keypair.ex
@@ -1,6 +1,6 @@
 defmodule Elixium.KeyPair do
   alias Elixium.Utilities
-  alias Elixium.Advanced
+  alias Elixium.Mnemonic
   use Bitwise
   require Integer
 
@@ -40,17 +40,17 @@ defmodule Elixium.KeyPair do
   """
   @spec create_mnemonic(binary) :: String.t()
   def create_mnemonic(private) do
-    Elixium.Advanced.from_entropy(private)
+    Mnemonic.from_entropy(private)
   end
 
   @doc """
-    Generates a keypair from the seed phrase
+    Generates a keypair from the seed phrase or from the private key, leading " " will switch to mnemonic to import key from
   """
   @spec gen_keypair(String.t()) :: {binary, binary}
   def gen_keypair(phrase) do
     case String.contains?(phrase, " ") do
       true ->
-        private = Elixium.Advanced.to_entropy(phrase)
+        private = Mnemonic.to_entropy(phrase)
         keys = get_from_private(private)
       false ->
         keys = get_from_private(phrase)
